Add files via upload
This file has functions for all the possible actions that we might need for the data purposes. It has function to read from the csv for a specific company, get dataset for required window sizes and horizons, split data into training, validation, and testing, partition data if you want to try that. I have left the print statements in so that you can test and visualise every step of the way, if you want to.